nvidia_docker/test/requirements.txt
tensorflow-gpu==2.5.0
scipy
